requirements.txt
Django
django_heroku
gunicorn
Keras
Pillow
tensorflow
whitenoise
django-crispy-forms
env-file
cloudinary
